pyproject.toml
[project]
name = "pps"
version = "0.1.0"
author = [
    {name = "ReactiveVector", email = "reactivevector@gmail.com"}
]
description = "Python-Project-Sample"
readme = "README.md"
requires-python = ">=3.10"
classifier = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Indenpendent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]


dependencies = [
    "flask==3.0.2",
    "mongo>=0.2.0",
    "pre-commit>=4.1.0",
    "pydantic>=2.10.6",
    "redis>=5.2.1",
    "ruff>=0.9.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pps"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",
    "F",
    "UP",
    "B",
    "SIM",
    "I",
]

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true
